Create votes

+++ db/migrate/20180401030004_create_votes.rb
@@ -3,7 +3,7 @@ class CreateVotes < ActiveRecord::Migration[5.1]
     create_table :votes do |t|
       t.references :product, index: true, foreign_key: true, null: false
       t.references :user, index: true, foreign_key: true, null: false
-      t.integer :star
+      t.float :star
       t.text :description
       t.datetime :deleted_at
